move HO 3D files to proper location, refactor includes
